+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.0.2
-
++ Added more safety checks for player info comparison 
++ Added a config option to delay the team syncing action
 
 # 1.0.1
 + Safe check for missing players when accepting invites